price type change

+++ src/main/java/com/project/oag/app/dto/ArtworkRequestDto.java
@@ -3,6 +3,7 @@ package com.project.oag.app.dto;
 import lombok.Data;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -10,7 +11,7 @@ public class ArtworkRequestDto {
 	private String artworkName;
 	private String artworkDescription;
 	private String artworkCategory;
-	private int price;
+	private BigDecimal price;
 	private String size;
 	List<MultipartFile> imageFiles;
 }
